renamed common_flows to lib, added all tests to the configmap, changed selection of locustfile to the prod simulator
