+++ kustomization.yaml
@@ -5,22 +5,36 @@ resources:
 - locust.yaml
 - fake-server.yaml
 configMapGenerator:
-- name: login-sign-up-sign-in-locustfile
+- name: login-locustfiles
   namespace: locust
   files:
+  - load_testing/ial2_sign_in.locustfile.py
+  - load_testing/ial2_sign_up.locustfile.py
+  - load_testing/prod_simulator.locustfile.py
+  - load_testing/sign_in.locustfile.py
+  - load_testing/sign_in_failure.locustfile.py
+  - load_testing/sign_in_remember_me.locustfile.py
+  - load_testing/sign_up.locustfile.py
   - load_testing/sign_up_sign_in.locustfile.py
+  - load_testing/sp_ial2_sign_in.locustfile.py
+  - load_testing/sp_ial2_sign_in_async.locustfile.py
+  - load_testing/sp_ial2_sign_up.locustfile.py
+  - load_testing/sp_ial2_sign_up_async.locustfile.py
+  - load_testing/sp_sign_in.locustfile.py
+  - load_testing/sp_sign_up.locustfile.py
+
 - name: login-loadtest-lib
   namespace: locust
   files:
-  - load_testing/common_flows/flow_helper.py
-  - load_testing/common_flows/flow_ial2_proofing.py
-  - load_testing/common_flows/flow_sign_in.py
-  - load_testing/common_flows/flow_sign_up.py
-  - load_testing/common_flows/flow_sp_ial2_sign_in.py
-  - load_testing/common_flows/flow_sp_ial2_sign_in_async.py
-  - load_testing/common_flows/flow_sp_ial2_sign_up.py
-  - load_testing/common_flows/flow_sp_ial2_sign_up_async.py
-  - load_testing/common_flows/flow_sp_sign_in.py
-  - load_testing/common_flows/flow_sp_sign_up.py
+  - load_testing/lib/flow_helper.py
+  - load_testing/lib/flow_ial2_proofing.py
+  - load_testing/lib/flow_sign_in.py
+  - load_testing/lib/flow_sign_up.py
+  - load_testing/lib/flow_sp_ial2_sign_in.py
+  - load_testing/lib/flow_sp_ial2_sign_in_async.py
+  - load_testing/lib/flow_sp_ial2_sign_up.py
+  - load_testing/lib/flow_sp_ial2_sign_up_async.py
+  - load_testing/lib/flow_sp_sign_in.py
+  - load_testing/lib/flow_sp_sign_up.py
 generatorOptions:
  disableNameSuffixHash: true
