+++ load_testing/sp_ial2_sign_in_async.locustfile.py
@@ -1,5 +1,5 @@
 from locust import HttpUser, TaskSet, task, between
-from common_flows import flow_sp_ial2_sign_in_async, flow_helper
+from lib import flow_sp_ial2_sign_in_async, flow_helper
 import logging
